+++ RSS/DTOs/CreateGameDTO.cs
@@ -6,7 +6,7 @@ namespace RSS.DTOs
         public string GameType { get; set; }
         public bool IsOpen { get; set; }
         public int PlayerCount { get; set; }
-        public int PricePerSquare { get; set; }
+        public decimal PricePerSquare { get; set; }
         public int SquareSelectionLimit { get; set; }
         public bool IsTurnBased { get; set; }
         public int TurnTimeoutSeconds { get; set; } = 60;
